refactor(logging): adjust hook registration
Use a thread safe sync.Once instead of a bool.
