+++ internal/logging/printf.go
@@ -27,3 +27,19 @@ type CtxPrintfLogger struct {
 func (l *CtxPrintfLogger) Printf(_ context.Context, format string, args ...any) {
 	l.logrus.Logf(l.level, format, args...)
 }
+
+// LoggerPrintf returns a new PrintfLogger given a level.
+func LoggerPrintf(level logrus.Level) (logger *PrintfLogger) {
+	return &PrintfLogger{
+		level:  level,
+		logrus: logrus.StandardLogger(),
+	}
+}
+
+// LoggerCtxPrintf returns a new CtxPrintfLogger given a level.
+func LoggerCtxPrintf(level logrus.Level) (logger *CtxPrintfLogger) {
+	return &CtxPrintfLogger{
+		level:  level,
+		logrus: logrus.StandardLogger(),
+	}
+}
